feat(reader): migrate ColumnSnap to commonMain; add iOS XCTests + scenario docs (#955)
- Move ColumnSnap object from app/androidMain to feature/reader/commonMain,
  replacing org.json (JSONObject.quote, JSONArray) and java.net.URLDecoder with
  pure-Kotlin equivalents (kotlinx.serialization.json + percentDecodeUtf8()).
- Promote NOTE_GLYPH_DECORATION_GROUP and NOTE_GLYPH_FOCUS_ID_JS_KEY to
  commonMain top-level constants; update all Android callers to import from
  the new package.
- Leave a thin `typealias ColumnSnap` shim in the Android app package so
  existing call sites compile without changes.
- Add ColumnSnapJsBuilderTest (commonTest, 29 cases) covering parseNarratedColumnsResult,
  navTargetFragmentId, and all JS builder shapes.
- Add AutoFollowJsTests.swift (7 WKWebView XCTest scenarios) exercising
  ColumnSnap.shared.autoFollowSnapJs().
- Add iOS scenario docs:
  - 08-auto-follow-js.md (12 scenarios; 7 implemented, 5 deferred with gap reason)
  - 08-epub-cfi-translator.md (pure-CFI in commonTest; file-reading gap)
  - 08-epub-search-service.md (all deferred; ReadiumSwiftNavigator.search() stub)
  - 08-scroll-boundary-navigation.md (all deferred; scroll mode not on iOS)
  - 08-annotation-focus-harness.md (all deferred; annotation focus not wired on iOS)

Closes #955
