+++ app/src/test/kotlin/com/riffle/app/feature/reader/NarratedColumnsResultParserTest.kt
@@ -1,5 +1,6 @@
 package com.riffle.app.feature.reader
 
+import com.riffle.feature.reader.ColumnSnap
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
