+++ docs/testing/ios-scenarios/08-auto-follow-js.md
@@ -1,0 +1,115 @@
+# Scenario 08: Auto-Follow JS (Readaloud page tracking)
+
+**Android reference:** `AutoFollowJsTest.kt` (12 tests)  
+**iOS implementation:** `AutoFollowJsTests.swift` (7 scenarios)  
+**KMP source:** `feature/reader/src/commonMain/kotlin/com/riffle/feature/reader/ColumnSnap.kt`  
+**Method:** `ColumnSnap.autoFollowSnapJs(text:)`
+
+---
+
+## Context
+
+The auto-follow probe runs on each narrated-sentence change during Readaloud playback.
+It locates the sentence by its full text (Readium strips media-overlay span ids), then:
+
+- **Scroll mode** (doc taller than viewport): centres the sentence vertically when it drifts
+  outside the middle half-viewport comfort band.
+- **Paginated mode** (doc fits viewport horizontally): floors `scrollLeft` to the column grid
+  that contains the sentence's start — a no-op if already on-page, a drift-correction if
+  off-grid, or a real page turn if the narration has crossed into another column.
+
+Returns `"on"` (sentence found and scroll/position updated or confirmed), or `"off"` (sentence
+not on this resource → caller must `go()` to the chapter that contains it).
+
+The JS is generated by `ColumnSnap.shared.autoFollowSnapJs(text:)` which is a shared KMP object.
+Its implementation is **identical** across Android (Android WebView) and iOS (WKWebView).
+
+---
+
+## Scenarios
+
+### 08-A: Scroll mode centres the narrated sentence
+- Load a tall document (height 6000 px) with the target sentence at top ~3000 px.
+- Evaluate `autoFollowSnapJs(targetText)`.
+- Assert result = `"on"`.
+- Assert `getBoundingClientRect().center.y` is within 12 px of `window.innerHeight / 2`.
+- **iOS:** `testScrollModeCentersTheNarratedSentence`
+
+### 08-B: Scroll mode does not bounce between two adjacent visible sentences
+- Load a tall document; scroll so both adjacent sentences A and B are inside the middle half.
+- Evaluate `autoFollowSnapJs(A)`, then `autoFollowSnapJs(B)`, then `autoFollowSnapJs(A)` again.
+- Assert `window.scrollY` deviation is ≤ 8 px for all three calls.
+- **iOS:** deferred — requires multiple sequential WKWebView.evaluateJavaScript calls and
+  precise viewport pre-positioning. Gap recorded; covered by Android `AutoFollowJsTest`.
+
+### 08-C: Scroll mode never scrolls horizontally
+- Load a tall document. Evaluate `autoFollowSnapJs(targetText)`.
+- Assert `window.scrollX == 0`.
+- **iOS:** `testScrollModeNeverScrollsHorizontally`
+
+### 08-D: Paginated returns `"on"` for a visible sentence without scrolling
+- Load a short (viewport-height) document. Evaluate `autoFollowSnapJs(onPageText)`.
+- Assert result = `"on"`.
+- **iOS:** `testPaginatedReturnsOnForVisibleSentence`
+
+### 08-E: Paginated snaps a visible sentence to its grid-aligned column
+- Load a short document. Set `scrollLeft = 10` (off-grid). Evaluate `autoFollowSnapJs(onPageText)`.
+- Assert result = `"on"` and `scrollX == 0`.
+- **iOS:** deferred — requires setting initial `scrollLeft` via a JS call before the probe.
+  Gap recorded; covered by Android `AutoFollowJsTest`.
+
+### 08-F: Paginated follows symmetrically (forward OR back)
+- Load a short document. Set `scrollLeft = 40`. Evaluate `autoFollowSnapJs(onPageText)`.
+- Assert result = `"on"` and `scrollX == 0`.
+- **iOS:** deferred — covered by Android test.
+
+### 08-G: Paginated snaps to the column containing the sentence
+- Load a short document with off-page sentence at `left:5000px`.
+- Evaluate `autoFollowSnapJs(offRightText)`.
+- Assert result = `"on"` and `scrollX % innerWidth == 0` (on the column grid).
+- **iOS:** deferred — snap to a column far off-screen requires precise layout measurement.
+
+### 08-H: Returns `"off"` for text not on the page
+- Evaluate `autoFollowSnapJs("Zzz nonexistent sentence text")`.
+- Assert result = `"off"`.
+- **iOS:** `testReturnsOffForTextNotOnPage`
+
+### 08-I: Empty text disables the probe
+- Evaluate `autoFollowSnapJs("")`.
+- Assert result = `"off"`.
+- **iOS:** `testEmptyTextDisablesTheProbe`
+
+### 08-J: Dateline collision does not false-match the previous chapter's dateline
+- Load a document containing `"LOG ENTRY: SOL 37"` (old chapter's dateline).
+- Evaluate `autoFollowSnapJs("LOG ENTRY: SOL 38")` (new chapter's dateline, not on this page).
+- Assert result = `"off"` (whole-sentence matching must not confuse the two).
+- **iOS:** `testDatelineCollisionDoesNotFalseMatchPreviousChapter`
+
+### 08-K: Dateline present on its own page still follows
+- Same document. Evaluate `autoFollowSnapJs("LOG ENTRY: SOL 37")`.
+- Assert result = `"on"`.
+- **iOS:** `testDatelinePresentOnItsOwnPageStillFollows`
+
+### 08-L: Follows a sentence split across inline markup nodes
+- Load a document where the target sentence is `Once we got <em>Hermes</em> moving we coasted`.
+- Evaluate `autoFollowSnapJs("Once we got Hermes moving we coasted")`.
+- Assert result = `"on"`.
+- **iOS:** `testFollowsSentenceSplitAcrossInlineMarkupNodes`
+
+---
+
+## Implementation notes
+
+`ColumnSnap` was migrated to `feature/reader/commonMain` in the same PR as this scenario doc,
+replacing `org.json.JSONObject.quote()` with `kotlinx.serialization.json.JsonPrimitive` and
+`java.net.URLDecoder` with a pure-Kotlin percent-decoder. The KMP object is exposed in the
+`Riffle` framework as `ColumnSnap.shared`. iOS tests call
+`ColumnSnap.shared.autoFollowSnapJs(text:)` and inject the result into a `WKWebView`.
+
+### Deferred scenarios
+
+Scenarios 08-E, 08-F, 08-G, 08-B require pre-positioning the scroll via JS before the probe
+and then asserting exact scroll values. The iOS XCTest infrastructure supports this but the
+column-grid math depends on `window.innerWidth` which in WKWebView headless tests may not
+match an exact pixel multiple. These are covered by the Android instrumented tests; iOS
+coverage is deferred until there is a real device in CI.
